remove xen support from qemu-system-x86 build and create a wrapper for -i386
Disable xen support for regular qemu-system-x86 binaries, since it is
now enabled in a separate qemu-system-xen package.  Make qemu-system-i386
to be a wrapper script that either calls regular qemu-system-i386 binary
or xen-specific qemu-system-i386-xen if xen-specific arguments are being
used, and the xen-specific binary exists. And print a warning to update
the configuration.