reimplement qemu-system installation in d/rules (remove d/inst-system)
This can be done in d/rules directly (with a mix of make and sh),
and it is consistent with linux-user install as well.

Also rearrange ${provides}/${breaks}/${replaces} variables in
qemu-system-foo packages for the kvm case to be more clear
what is going on, and fix a bug in there.  Rename the variables
to qemu:provides and qemu:kvm.