specify version for qemu-kvm in Provides, also stop Breaking/Replacing qemu-kvm
The Break/Replaces for qemu-kvm was there to replace old non-virtual
qemu-kvm package. It does not exist for a few releases, so this now
can be dropped.  Which is handy, because now with version string for
qemu-kvm in Provides, we'd need two separate variables to hold
versioned and unversioned qemu-kvm to refer to them in d/control.
Now whole ${qemu:Provides} is constructed in d/rules:install-system.