d/control: unify and simplify arch strings for --enable-foo and dependencies
Do not list the linux- part in linux-foo since this is often forgotten.
This brings us many forgotten --enable-foo, such as --enable-spice on
linux-x32, since linux- part has been missing in there and old extract-config-opts
did not work with os-less arch strings.

Now [dependencies] have the same contents as --enable-foo, and is
written in short way for default linux os.