d/control: generate Architecture: field dynamically from d/rules
This is to enusre that everything is consistent (groups of packages
needs to have the same Architecture value), and to be able to
restrict dependencies to system or user architectures in the future.

Use the same simile mechanism which is already used for vendor-specific
configuration, by replacing :system-arch-linux:, :system-arch:, :user-arch:
and :utils-arch: in d/configure-in.