d/rules: move qemu-user-binfmt doclink to the proper place, and remove overrides
we had dh_installdocs and dh_installchangelogs overrides just in order
to omit them for package(s) where the doc dir is a symlink to another
package's doc dir. But both are smart enough to not act on packages
for which dh_installdocs --link-doc= has been called before. So just
remove the overrides.