Revert "stop using custom $argv[0] for binfmt"
This reverts commit 1f805a1389b80ecda82a07b4ec190a0352fa3eb0.

Re-rely on qemu-user's argv0 to detect it is running in binfmt context.
The problem is that while we ship kernel which can pass this info the
qemu way, there are many containers which are running on older kernels
still, including bullseye kernel (5.10) which does not have this feature.
Keep it for a bit more.

Resolve conflicts in d/binfmt-install.