d/rules: split out qemu-user build out out of main qemu build
when both system and linux-user builds are enabled, linux-user
build is getting features only relevant for system (softmmu)
configuration, like linking with liburing, libnuma and other
softmmu-only stuff.  So build it separately.

This not only makes qemu-user smaller and neater, but it also
makes Built-Using field for qemu-user-static (which is generated
from Depends field of qemu-user) accurate.